Refactor: refactor question card component to support swappable and non-swappable state - convert state by allowSwap flag
part of SWM-97
related to #23
